#935 fixed (Modelica.Blocks.Math.RootMeanSquare might fail):
Added a parameter in block "Mean" to state, if the output is guaranteed to be >= 0.0.
In such a case, the output of block "Mean" is computed as max(0, y_computed), i.e., it is
guaranteed to be >= 0.0. In the RootMeanSquare block, this parameter is appropriately set and
then it is guaranteed that the input to the Sqrt block is >= 0.0.

Additionally, added a test model in ModelicaTest to check the desired behavior.



git-svn-id: https://svn.modelica.org/projects/Modelica/trunk@5860 7ce873d0-865f-4ce7-a662-4bb36ea78beb